requirements.txt
2to3==1.0
aiohttp==3.6.2
appdirs==1.4.4
astroid==2.3.3
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cssselect==1.1.0
cycler==0.10.0
discord.py==1.4.1
fake-useragent==0.1.11
idna==2.8
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.5.2
matplotlib==3.1.3
mccabe==0.6.1
multidict==4.7.6
mysql-connector-python==8.0.18
noise==1.2.2
numpy==1.18.1
pandas==1.0.1
parse==1.18.0
pyee==7.0.3
pylint==2.4.4
pyparsing==2.4.6
pyppeteer==0.2.2
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-html==0.10.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0.1
tdk==1.2.0
tqdm==4.49.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
w3lib==1.22.0
websockets==8.1
wikipedia==1.4.0
wrapt==1.11.2
yarl==1.5.1